src/Batch-dbt/requirements.txt
pandas
google-cloud-bigquery
Flask==3.0.0